Specify a Tutorial model
A record of the Tutorial model should have the following fields [datatypes]:
  name [string]
  slug [string]

The Tutorial model should validate for the presence of each field before creating a new record.

Since all of the attributes are of type string, and ActiveRecord automatically will change any non-string value into a string before saving it in the database, there isn't a need to validate for input types.