Install factory_bot_rails for generating test data
This commit installs factory_bot_rails and configures it by making the FactoryBot methods available in the RSpec specs.

This will be useful for generating test data. For example, in testing the Feedback model, factory_bot_rails can be used to generate a mock Tutorial instance to use for associations. This will allow testing that a Feedback belongs to a Tutorial, without relying on the actual implementation of the Tutorial model itself.